Move BoostCardInfo to data/Uniboost.ts
